Extract shared DriveDoc (clean_title/date-stamps/coerce) from DriveSource
